HOME-89 - Consider code refactor to be more model oriented
