+++ datasources/persistence/persistence.go
@@ -5,6 +5,11 @@ import (
     "gopkg.in/mgo.v2"
 )
 
+// IPersistance - interface for user settings and general purpose storage
+type IPersistance interface {
+    GetCollection(string) *mgo.Collection
+}
+
 // Persistance - data source keeping system state and user data
 type Persistance struct {
     session *mgo.Session
